Moved async article recommendation interfaces
