+++ sciety_labs/app/app_providers_and_models.py
@@ -17,7 +17,7 @@ from sciety_labs.providers.interfaces.article_recommendation import (
     ArticleRecommendationProvider,
     SingleArticleRecommendationProvider
 )
-from sciety_labs.providers.async_article_recommendation import (
+from sciety_labs.providers.interfaces.async_article_recommendation import (
     AsyncSingleArticleRecommendationProvider
 )
 from sciety_labs.providers.async_crossref import (
