+++ sciety_labs/providers/opensearch/async_opensearch_article_recommendation.py
@@ -12,7 +12,7 @@ from sciety_labs.providers.interfaces.article_recommendation import (
     ArticleRecommendationFilterParameters,
     ArticleRecommendationList
 )
-from sciety_labs.providers.async_article_recommendation import (
+from sciety_labs.providers.interfaces.async_article_recommendation import (
     AsyncSingleArticleRecommendationProvider
 )
 from sciety_labs.providers.async_crossref import AsyncCrossrefMetaDataProvider
